+++ src/scholarag/app/routers/retrieval.py
@@ -90,6 +90,11 @@ async def retrieval(
     -------
         A list of article titles and paragraphs.
     """  # noqa: D301, D400, D205
+    if len(request.query) > settings.misc.query_max_size:
+        raise HTTPException(
+            status_code=413,
+            detail=f"Query string has {len(request.query)} characters. Maximum allowed is {settings.misc.query_max_size}.",
+        )
     start = time.time()
     logger.info("Finding documents ...")