+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Fixed
+- Limit query size.
 - Add OBI copyright.
 
 ## [0.0.8] - 10.12.2024